feat: Implement core frontend components, rewrote the whole front end codebase from scratch. whole structure is different now
