+++ frontend/components/analysis/analysis-dashboard.tsx
@@ -7,11 +7,16 @@ import { AnalysisResult, TempoPoint, TranscriptWord, LongPause } from "@/types/a
 import { resolveVideoUrl, getPdfUrl } from "@/lib/api";
 
 import { cn } from "@/lib/utils";
-import { SmoothScroll } from "./smooth-scroll";
-import { TempoChart } from "./analysis/tempo-chart";
-import { ComingSoonNotification } from "./coming-soon-notification";
-import { PdfExportDropdown } from "./pdf-export-modal";
-import { VideoPlayer, VideoPlayerRef } from "./video-player";
+import { SmoothScroll } from "@/components/ui/smooth-scroll";
+import { TempoChart } from "@/components/analysis/tempo-chart";
+import { ComingSoonNotification } from "@/components/shared/coming-soon-notification";
+import { PdfExportDropdown } from "@/components/shared/pdf-export-modal";
+import { VideoPlayer, VideoPlayerRef } from "@/components/shared/video-player";
+import { IconTarget, IconDoc, IconBolt, IconClock, IconHand } from "@/components/ui/icons";
+import { StatBadge } from "@/components/ui/stat-badge";
+import { QuickStat } from "@/components/ui/quick-stat";
+import { InsightCard } from "@/components/analysis/insight-card";
+import { AnimatedConfidenceGauge } from "@/components/analysis/animated-confidence-gauge";
 
 type Props = {
   result: AnalysisResult;
@@ -28,40 +33,6 @@ type ViewportRect = {
 const easeOutExpo = (t: number) => (t === 1 ? 1 : 1 - Math.pow(2, -10 * t));
 const easeOutCubic = (t: number) => 1 - Math.pow(1 - t, 3);
 const easeOutQuint = (t: number) => 1 - Math.pow(1 - t, 5);
-
-const IconTarget = ({ className }: { className?: string }) => (
-  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
-    <circle cx="12" cy="12" r="9" opacity="0.5" />
-    <circle cx="12" cy="12" r="4" />
-    <path d="M12 3v3M12 18v3M3 12h3M18 12h3" />
-  </svg>
-);
-const IconDoc = ({ className }: { className?: string }) => (
-  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
-    <path d="M7 3h7l4 4v14H7z" />
-    <path d="M14 3v5h5" opacity="0.6" />
-    <path d="M9 13h6M9 17h6" opacity="0.7" />
-  </svg>
-);
-const IconBolt = ({ className }: { className?: string }) => (
-  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
-    <path d="M13 2 6 13h5l-1 9 7-11h-5z" />
-  </svg>
-);
-const IconClock = ({ className }: { className?: string }) => (
-  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
-    <circle cx="12" cy="12" r="9" />
-    <path d="M12 7v6l4 2" />
-  </svg>
-);
-const IconHand = ({ className }: { className?: string }) => (
-  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
-    <path d="M18 11v-1a2 2 0 0 0-2-2h-1" />
-    <path d="M14 10V9a2 2 0 0 0-2-2h-1" />
-    <path d="M10 9.5V4a2 2 0 0 0-2-2v0a2 2 0 0 0-2 2v7" />
-    <path d="M18 11a2 2 0 1 1 4 0v3a8 8 0 0 1-8 8h-1a8 8 0 0 1-8-8 2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2a2 2 0 0 0 2 2h2a2 2 0 0 0 2-2v-3z" />
-  </svg>
-);
 
 export function AnalysisDashboard({ result, onBack }: Props) {
   const playerRef = useRef<VideoPlayerRef>(null);
@@ -671,20 +642,6 @@ function formatTime(seconds: number): string {
   return `${m}:${s.toString().padStart(2, "0")}`;
 }
 
-function StatBadge({ label, value, accent }: { label: string; value: string; accent?: boolean }) {
-  return (
-    <div
-      className={cn(
-        "hidden rounded-full px-3 py-1.5 text-xs sm:block",
-        accent ? "bg-white/15 font-medium" : "bg-white/5"
-      )}
-    >
-      <span className="text-white/50">{label}:</span>{" "}
-      <span className="font-semibold">{value}</span>
-    </div>
-  );
-}
-
 function TabButton({
   active,
   onClick,
@@ -706,399 +663,6 @@ function TabButton({
     >
       {children}
     </button>
-  );
-}
-
-function QuickStat({
-  icon,
-  label,
-  value,
-}: {
-  icon: React.ReactNode;
-  label: string;
-  value: string;
-}) {
-  return (
-    <div className="group relative flex flex-col gap-2 rounded-3xl border border-white/10 bg-black/40 backdrop-blur-2xl p-5 transition-all duration-500 hover:border-white/20 hover:bg-white/5 shadow-[0_8px_32px_0_rgba(0,0,0,0.3)] hover:shadow-[0_12px_40px_rgba(255,255,255,0.05)]">
-      <div className="absolute inset-0 bg-gradient-to-br from-white/5 to-transparent pointer-events-none rounded-3xl" />
-      <div className="flex items-center gap-4 relative z-10 w-full overflow-hidden">
-        <div className="flex h-12 w-12 shrink-0 items-center justify-center rounded-2xl bg-white/5 shadow-inner border border-white/10 transition-transform duration-500 group-hover:scale-110 group-hover:bg-white/10">
-          <span className="drop-shadow-md">{icon}</span>
-        </div>
-        <div className="min-w-0 flex-1">
-          <p className="text-[10px] font-semibold text-white/40 tracking-widest uppercase truncate">{label}</p>
-          <p className="text-[22px] font-medium tabular-nums tracking-tight mt-0.5 text-white drop-shadow-sm">{value}</p>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function AnimatedTempoChart({ data, mounted }: { data: TempoPoint[]; mounted: boolean }) {
-  const [animProgress, setAnimProgress] = useState(0);
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const [hovered, setHovered] = useState<TempoPoint | null>(null);
-  const [hoverPos, setHoverPos] = useState<{ x: number; y: number }>({ x: 0, y: 0 });
-
-  const maxWpm = useMemo(() => Math.max(...data.map((d) => d.wpm), 1), [data]);
-  const maxTime = useMemo(() => Math.max(...data.map((d) => d.time), 1), [data]);
-
-  useEffect(() => {
-    if (!mounted) return;
-    let start = Date.now();
-    const duration = 1200;
-    const animate = () => {
-      const elapsed = Date.now() - start;
-      const progress = Math.min(elapsed / duration, 1);
-      setAnimProgress(1 - Math.pow(1 - progress, 3));
-      if (progress < 1) requestAnimationFrame(animate);
-    };
-    const timer = setTimeout(animate, 400);
-    return () => clearTimeout(timer);
-  }, [mounted]);
-
-  useEffect(() => {
-    const canvas = canvasRef.current;
-    if (!canvas || !data.length) return;
-    const ctx = canvas.getContext("2d");
-    if (!ctx) return;
-
-    const dpr = window.devicePixelRatio || 1;
-    const w = 360;
-    const h = 200;
-    canvas.width = w * dpr;
-    canvas.height = h * dpr;
-    canvas.style.width = `${w}px`;
-    canvas.style.height = `${h}px`;
-    ctx.scale(dpr, dpr);
-
-    const pad = 30;
-
-    ctx.clearRect(0, 0, w, h);
-
-    ctx.strokeStyle = "rgba(255,255,255,0.06)";
-    ctx.lineWidth = 1;
-    for (let i = 0; i <= 5; i++) {
-      const y = pad + ((h - pad * 2) / 5) * i;
-      ctx.beginPath();
-      ctx.moveTo(pad, y);
-      ctx.lineTo(w - pad, y);
-      ctx.stroke();
-    }
-
-    const visibleCount = Math.floor(data.length * animProgress);
-    const toPoint = (d: TempoPoint) => {
-      const x = pad + (d.time / maxTime) * (w - pad * 2);
-      const y = h - pad - (d.wpm / maxWpm) * (h - pad * 2);
-      return { x, y };
-    };
-
-    if (visibleCount > 0) {
-      ctx.beginPath();
-      for (let i = 0; i <= visibleCount; i++) {
-        const d = data[i];
-        if (!d) continue;
-        const { x, y } = toPoint(d);
-        if (i === 0) {
-          ctx.moveTo(x, y);
-        } else {
-          const prev = data[i - 1];
-          const { x: px, y: py } = toPoint(prev);
-          const cx = (px + x) / 2;
-          ctx.quadraticCurveTo(cx, py, x, y);
-        }
-      }
-      const last = data[Math.min(visibleCount, data.length - 1)];
-      if (last) {
-        const { x } = toPoint(last);
-        ctx.lineTo(x, h - pad);
-        ctx.lineTo(pad, h - pad);
-        ctx.closePath();
-        const grad = ctx.createLinearGradient(0, pad, 0, h - pad);
-        grad.addColorStop(0, "rgba(255,255,255,0.18)");
-        grad.addColorStop(1, "rgba(255,255,255,0.02)");
-        ctx.fillStyle = grad;
-        ctx.fill();
-      }
-    }
-
-    ctx.beginPath();
-    ctx.lineWidth = 2.2;
-    ctx.strokeStyle = "rgba(255,255,255,0.9)";
-    ctx.lineCap = "round";
-    ctx.lineJoin = "round";
-    for (let i = 0; i <= visibleCount; i++) {
-      const d = data[i];
-      if (!d) continue;
-      const { x, y } = toPoint(d);
-      if (i === 0) ctx.moveTo(x, y);
-      else {
-        const prev = data[i - 1];
-        const { x: px, y: py } = toPoint(prev);
-        const cx = (px + x) / 2;
-        ctx.quadraticCurveTo(cx, py, x, y);
-      }
-    }
-    ctx.stroke();
-
-    ctx.save();
-    ctx.strokeStyle = "rgba(255,255,255,0.35)";
-    ctx.lineWidth = 6;
-    ctx.filter = "blur(6px)";
-    ctx.stroke();
-    ctx.restore();
-
-    for (let i = 0; i <= visibleCount; i++) {
-      const d = data[i];
-      if (!d) continue;
-      const { x, y } = toPoint(d);
-      ctx.beginPath();
-      ctx.arc(x, y, 4.5, 0, Math.PI * 2);
-      ctx.fillStyle = d.zone === "red" ? "#f87171" : d.zone === "yellow" ? "#fbbf24" : "#fff";
-      ctx.fill();
-    }
-
-    ctx.fillStyle = "rgba(255,255,255,0.45)";
-    ctx.font = "10px sans-serif";
-    ctx.fillText("0", pad - 12, h - pad + 14);
-    ctx.fillText(formatTime(maxTime), w - pad - 16, h - pad + 14);
-    ctx.fillText(`${maxWpm.toFixed(0)} wpm`, pad - 12, pad + 4);
-  }, [data, maxWpm, maxTime, animProgress]);
-
-  return (
-    <div
-      className={cn(
-        "relative overflow-hidden rounded-3xl border border-white/10 bg-black/40 backdrop-blur-2xl p-6 transition-all duration-500 shadow-[0_8px_32px_0_rgba(0,0,0,0.3)] hover:shadow-[0_12px_40px_rgba(255,255,255,0.05)] hover:border-white/20",
-        mounted ? "translate-y-0 opacity-100" : "translate-y-4 opacity-0"
-      )}
-      style={{ transitionDelay: "500ms" }}
-    >
-      <div className="absolute inset-0 bg-gradient-to-br from-white/5 to-transparent pointer-events-none" />
-      <div className="relative z-10">
-        <div className="mb-4 flex items-center justify-between">
-          <div className="space-y-0.5">
-            <h3 className="text-sm font-semibold text-white/90">Темп речи</h3>
-            <p className="text-[11px] text-white/50 tracking-wide uppercase">Гладкая линия</p>
-          </div>
-          <span className="text-xs font-medium text-white/60 bg-white/10 px-3 py-1.5 rounded-full border border-white/10 drop-shadow-sm">max {maxWpm.toFixed(0)} wpm</span>
-        </div>
-        <div
-        className="relative"
-        onMouseLeave={() => setHovered(null)}
-        onMouseMove={(e) => {
-          const canvas = canvasRef.current;
-          if (!canvas) return;
-          const rect = canvas.getBoundingClientRect();
-          const x = e.clientX - rect.left;
-          const pad = 30;
-          const rel = (x - pad) / (rect.width - pad * 2);
-          const time = Math.max(0, Math.min(1, rel)) * maxTime;
-          let best: TempoPoint | null = null;
-          let min = Infinity;
-          for (const d of data) {
-            const dist = Math.abs(d.time - time);
-            if (dist < min) {
-              min = dist;
-              best = d;
-            }
-          }
-          if (best) {
-            setHovered(best);
-            setHoverPos({ x: e.clientX - rect.left + 12, y: e.clientY - rect.top - 10 });
-          }
-        }}
-      >
-        <canvas ref={canvasRef} style={{ width: "100%", height: 200 }} />
-        {hovered && (
-          <div
-            className="pointer-events-none absolute z-10 rounded-lg border border-white/10 bg-black/80 px-3 py-2 text-[11px] shadow-xl backdrop-blur-md"
-            style={{ left: hoverPos.x, top: hoverPos.y }}
-          >
-            <div className="font-semibold">t = {hovered.time.toFixed(1)}s</div>
-            <div className="text-white/70">WPM: {hovered.wpm.toFixed(1)}</div>
-            <div
-              className={cn(
-                "mt-1 inline-flex rounded-full px-2 py-0.5 text-[10px]",
-                hovered.zone === "red" && "bg-red-500/20 text-red-200",
-                hovered.zone === "yellow" && "bg-amber-500/20 text-amber-200",
-                hovered.zone !== "red" && hovered.zone !== "yellow" && "bg-white/10 text-white/70"
-              )}
-            >
-              {hovered.zone}
-            </div>
-          </div>
-        )}
-      </div>
-      </div>
-    </div>
-  );
-}
-
-function AnimatedConfidenceGauge({
-  total,
-  components,
-  mounted,
-}: {
-  total: number;
-  components: AnalysisResult["confidence_index"]["components"];
-  mounted: boolean;
-}) {
-  const [animValue, setAnimValue] = useState(0);
-  const normalizedTotal = Math.min(100, Math.max(0, total));
-
-  useEffect(() => {
-    if (!mounted) return;
-    let start = Date.now();
-    const duration = 1500;
-    const animate = () => {
-      const elapsed = Date.now() - start;
-      const progress = Math.min(elapsed / duration, 1);
-      const eased = 1 - Math.pow(1 - progress, 4);
-      setAnimValue(normalizedTotal * eased);
-      if (progress < 1) requestAnimationFrame(animate);
-    };
-    const timer = setTimeout(animate, 600);
-    return () => clearTimeout(timer);
-  }, [mounted, normalizedTotal]);
-
-  const radius = 45;
-  const circumference = 2 * Math.PI * radius;
-  const offset = circumference * (1 - animValue / 100);
-
-  return (
-    <div className="relative overflow-hidden rounded-3xl border border-white/10 bg-black/40 backdrop-blur-2xl p-6 shadow-[0_8px_32px_0_rgba(0,0,0,0.3)] transition-all duration-500 hover:border-white/20">
-      <div className="absolute inset-0 bg-gradient-to-br from-white/5 to-transparent pointer-events-none" />
-      <div className="relative z-10">
-        <h3 className="text-[10px] font-semibold text-white/40 uppercase tracking-widest mb-5">Детали уверенности</h3>
-        <div className="flex items-center gap-5">
-          <div className="relative h-20 w-20 flex-shrink-0 drop-shadow-[0_0_15px_rgba(255,199,0,0.3)]">
-            <svg className="h-full w-full -rotate-90" viewBox="0 0 100 100">
-              <circle cx="50" cy="50" r={radius} stroke="rgba(255,255,255,0.06)" strokeWidth="6" fill="none" />
-              <circle
-                cx="50"
-                cy="50"
-                r={radius}
-                stroke="#FFC700"
-                strokeWidth="6"
-                strokeDasharray={circumference}
-                strokeDashoffset={offset}
-                strokeLinecap="round"
-                fill="none"
-                style={{ transition: "stroke-dashoffset 0.2s linear" }}
-              />
-            </svg>
-            <div className="absolute inset-0 flex items-center justify-center">
-              <span className="text-2xl font-bold tabular-nums drop-shadow-md">{Math.round(animValue)}</span>
-            </div>
-          </div>
-          <div className="flex-1 space-y-4">
-            <MiniBar label="Громкость" value={Math.min(100, Math.max(0, components.volume_score))} delay={700} mounted={mounted} />
-            <MiniBar label="Чистота речи" value={Math.min(100, Math.max(0, components.filler_score))} delay={800} mounted={mounted} />
-            <MiniBar label="Взгляд" value={Math.min(100, Math.max(0, components.gaze_score))} delay={900} mounted={mounted} />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function MiniBar({
-  label,
-  value,
-  delay,
-  mounted,
-}: {
-  label: string;
-  value: number;
-  delay: number;
-  mounted: boolean;
-}) {
-  const [animValue, setAnimValue] = useState(0);
-
-  useEffect(() => {
-    if (!mounted) return;
-    const timer = setTimeout(() => {
-      let start = Date.now();
-      const duration = 800;
-      const animate = () => {
-        const elapsed = Date.now() - start;
-        const progress = Math.min(elapsed / duration, 1);
-        setAnimValue(value * (1 - Math.pow(1 - progress, 3)));
-        if (progress < 1) requestAnimationFrame(animate);
-      };
-      animate();
-    }, delay);
-    return () => clearTimeout(timer);
-  }, [mounted, value, delay]);
-
-  return (
-    <div>
-      <div className="flex justify-between text-[11px] mb-1">
-        <span className="text-white/40">{label}</span>
-        <span className="text-white/70 tabular-nums">{Math.round(animValue)}</span>
-      </div>
-      <div className="h-1 rounded-full bg-white/5 overflow-hidden">
-        <div
-          className="h-full rounded-full bg-white/30 transition-all duration-100"
-          style={{ width: `${Math.min(100, Math.max(0, animValue))}%` }}
-        />
-      </div>
-    </div>
-  );
-}
-
-function InsightCard({
-  title,
-  content,
-  delay,
-  mounted,
-  accent,
-  className,
-}: {
-  title: string;
-  content: string;
-  delay: number;
-  mounted: boolean;
-  accent?: "red" | "amber";
-  className?: string;
-}) {
-  return (
-    <div
-      className={cn(
-        "group relative overflow-hidden rounded-3xl border transition-all duration-500 shadow-[0_8px_32px_0_rgba(0,0,0,0.3)] hover:shadow-[0_12px_40px_rgba(255,255,255,0.05)]",
-        accent === "red" ? "border-rose-500/20 bg-black/40 backdrop-blur-2xl hover:border-rose-500/40" : 
-        accent === "amber" ? "border-amber-500/20 bg-black/40 backdrop-blur-2xl hover:border-amber-500/40" : 
-        "border-white/10 bg-black/40 backdrop-blur-2xl hover:border-white/30",
-        mounted ? "opacity-100 translate-y-0" : "opacity-0 translate-y-4",
-        className
-      )}
-    >
-      <div className="absolute inset-0 bg-gradient-to-br from-white/5 to-transparent pointer-events-none" />
-      <div className="relative z-10 p-6">
-        {/* Accent line */}
-        {accent && (
-          <div
-            className={cn(
-              "absolute left-0 top-0 bottom-0 w-1",
-              accent === "red" && "bg-rose-500/50 shadow-[0_0_10px_rgba(244,63,94,0.5)]",
-              accent === "amber" && "bg-amber-500/50 shadow-[0_0_10px_rgba(245,158,11,0.5)]"
-            )}
-          />
-        )}
-
-        <h3
-          className={cn(
-            "text-base font-semibold mb-3 tracking-tight drop-shadow-sm",
-            accent === "amber" && "text-amber-200/90",
-            accent === "red" && "text-rose-300/90",
-            !accent && "text-white"
-          )}
-        >
-          {title}
-        </h3>
-        <p className="text-[14px] leading-relaxed text-white/70 whitespace-pre-wrap">{content}</p>
-      </div>
-    </div>
   );
 }
