+++ frontend/components/landing/hero.tsx
