+++ frontend/components/shared/video-player.tsx
