+++ frontend/components/ui/icons.tsx
@@ -1,0 +1,39 @@
+import React from "react";
+
+export const IconTarget = ({ className }: { className?: string }) => (
+  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
+    <circle cx="12" cy="12" r="9" opacity="0.5" />
+    <circle cx="12" cy="12" r="4" />
+    <path d="M12 3v3M12 18v3M3 12h3M18 12h3" />
+  </svg>
+);
+
+export const IconDoc = ({ className }: { className?: string }) => (
+  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
+    <path d="M7 3h7l4 4v14H7z" />
+    <path d="M14 3v5h5" opacity="0.6" />
+    <path d="M9 13h6M9 17h6" opacity="0.7" />
+  </svg>
+);
+
+export const IconBolt = ({ className }: { className?: string }) => (
+  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
+    <path d="M13 2 6 13h5l-1 9 7-11h-5z" />
+  </svg>
+);
+
+export const IconClock = ({ className }: { className?: string }) => (
+  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
+    <circle cx="12" cy="12" r="9" />
+    <path d="M12 7v6l4 2" />
+  </svg>
+);
+
+export const IconHand = ({ className }: { className?: string }) => (
+  <svg className={className} width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.6">
+    <path d="M18 11v-1a2 2 0 0 0-2-2h-1" />
+    <path d="M14 10V9a2 2 0 0 0-2-2h-1" />
+    <path d="M10 9.5V4a2 2 0 0 0-2-2v0a2 2 0 0 0-2 2v7" />
+    <path d="M18 11a2 2 0 1 1 4 0v3a8 8 0 0 1-8 8h-1a8 8 0 0 1-8-8 2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2a2 2 0 0 0 2 2h2a2 2 0 0 0 2-2v-3z" />
+  </svg>
+);
