+++ frontend/components/upload/processing-overlay.tsx
